pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "neural_gfca"
authors = [
    { name = "National Institute of Advanced Industrial Science and Technology", email = "y.bando@aist.go.jp" },
]
classifiers = ["License :: OSI Approved :: MIT License"]
readme = "README.md"
license = {file = "LICENSE"}
description = "A library of Guided Neural FCA"
version = "0.0.1"

dependencies = [
    "aiaccel@git+https://github.com/aistairc/aiaccel.git@1a3360ce31d8fce80713d42807140ed3b69d0fe0",
    "einops>=0.8.0",
    "torch>=2.4.0",
    "lightning>=2.4.0",
    "numpy>=1.23.5",
    "matplotlib>=3.8.3",
    "soundfile",
    "hydra-core",
    "omegaconf",
    "torchaudio",
    "rich",
    "wpe-gpu",
]

[project.urls]
Home = "https://ybando.jp"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "UP", "B", "SIM", "I"]

[tool.ruff.lint.isort]
force-sort-within-sections = true

section-order = [
    "future",
    "typing",
    "standard-library",
    "utilities",
    "datascience",
    "torch",
    "torch-third-party",
    "third-party",
    "audio",
    "first-party",
    "local-folder"
]

[tool.ruff.lint.isort.sections]
"typing" = ["typing"]
"utilities" = ["progressbar", "omegaconf", "hydra"]
"datascience" = ["numpy", "scipy", "pandas", "matplotlib", "opt_einsum", "einops"]
"torch" = ["torch"]
"torch-third-party" = ["torchaudio", "torchvision", "auraloss", "lightning", "einops.layers"]
"audio" = ["librosa", "pypesq", "pystoi", "soundfile"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]